fix: Change return type of reduce operations to `Fold`
This changes all of the iterator reduce operations (`count`, `collect`,
etc) from `impl Future<Output = ...>` to use `Fold` or the new
`FoldWith` structures. The change is needed to give the ability to have
these operations have derivable traits like `Clone`, `Copy`, `Send`, and
`Sync` picked up automatically. With the `impl Future<Output = ...>`,
the returned opaque type can not pick up these traits without an
explicit `+ Send`, but there is no way to express that in the case where
the inputs (e.g.: `self`) do not have those traits.

Long-term, it might be preferable to have each `FallibleAsyncIterator`
operation have a dedicated structure for implementation. The type names
from function chains are a bit unwieldy (`Fold<FoldWith<Fold<Fold<...`)
and this is likely to get worse. The compiler might be able to help this
with some form of diagnostic emission.